requirements.txt
gql==3.0.0
pydot
python-dotenv
